requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.5
django-filter==2.4.0
django-heroku==0.3.1
django-model-utils==4.0.0
django-nested-admin==3.3.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-writable-nested==0.6.0
gunicorn==20.0.4
Pillow==7.1.1
psycopg2==2.8.6
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-monkey-business==1.0.0
pytz==2019.3
six==1.15.0
sqlparse==0.3.1
whitenoise==5.0.1
